crear rutas para equipo, reserva y mantenimiento
